refactor: split up Storage, add components of Green threads
